move MAX_LEGEND_ITEMS
